requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-cuser==2017.3.16
django-filter==22.1
django-import-export==3.0.2
django-templated-mail==1.1.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.2
modeltranslation==0.25
numpy==1.24.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.3
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
